+++ lib/ui/groups/GroupDetailPanel.dart
@@ -166,7 +166,11 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
       Groups.notifyGroupUpdated,
       Groups.notifyGroupEventsUpdated,
       Groups.notifyGroupPostsUpdated,
-      Polls.notifyCreated]);
+      Polls.notifyCreated,
+      Polls.notifyStatusChanged,
+      Polls.notifyVoteChanged,
+      Polls.notifyResultsChanged,
+    ]);
 
     _loadGroup(loadEvents: true);
   }
@@ -416,7 +420,11 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
       _refreshCurrentPosts(delta: param is int ? param : null);
     } else if (name == Polls.notifyCreated) {
       _refreshPolls();
-    } else if (name == AppLivecycle.notifyStateChanged) {
+    } else if (name == Polls.notifyVoteChanged
+            || name == Polls.notifyResultsChanged 
+            || name == Polls.notifyStatusChanged) {
+      _onPollUpdated(param); // Deep collection update single element (do not reload whole list)
+    }else if (name == AppLivecycle.notifyStateChanged) {
       _onAppLivecycleStateChanged(param);
     }
   }
@@ -1342,6 +1350,19 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CreatePollPanel(group: _group)));
   }
 
+  void _onPollUpdated(String? pollId) {
+    if(pollId!= null && _groupPolls!=null
+        && _groupPolls?.firstWhere((element) => pollId == element.pollId) != null) { //This is Group poll
+
+      Poll? poll = Polls().getPoll(pollId: pollId);
+      if (poll != null) {
+        setState(() {
+          _updatePollInList(poll);
+        });
+      }
+    }
+  }
+
   Future<void>_onPullToRefresh() async {
     Group? group = await Groups().loadGroup(widget.groupId); // The same as _refreshGroup(refreshEvents: true) but use await to show the pull to refresh progress indicator properly
     if ((group != null)) {
@@ -1405,6 +1426,17 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
 
   bool get _showMembershipBadge {
     return _isMemberOrAdmin || _isPending;
+  }
+
+  //Util
+  void _updatePollInList(Poll? poll) {
+    if ((poll != null) && (_groupPolls != null)) {
+      for (int index = 0; index < _groupPolls!.length; index++) {
+        if (_groupPolls![index].pollId == poll.pollId) {
+          _groupPolls![index] = poll;
+        }
+      }
+    }
   }
 }