+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Issues with Group Polls [#1157](https://github.com/rokwire/illinois-app/issues/1157).
+
 ## [3.1.9] - 2022-01-07
 ### Fixed
 - Loading Event images [#1184](https://github.com/rokwire/illinois-app/issues/1184).